+++ src/controllers/Controller.js
@@ -1,9 +1,9 @@
 import View from '../views/View.js';
 import { $, default as DOM } from '../utils/DOMUtils.js';
 import { default as DB } from '../model/database.js';
-import ProductAddManager from './ProductAddManager.js';
-import VendingMachineChargeManager from './VendingMachineChargeManager.js';
-import ProductPurchaseManager from './ProductPurchaseManager.js';
+import ProductAddManager from './managers/ProductAddManager.js';
+import VendingMachineChargeManager from './managers/VendingMachineChargeManager.js';
+import ProductPurchaseManager from './managers/ProductPurchaseManager.js';
 
 export default class Controller {
   constructor() {
